requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.0.0
psycopg2==2.9.6
gunicorn==23.0.0
python-dotenv==1.0.0
typing_extensions==4.12.2
packaging==24.1






